Make it bigger.
The key piece is that ``levelDistance: 175,`` line.  That's what
increases the distance between each level of the tree and makes it all
bigger. Everything else is just a css change to accomodate that.

The divs are now no-longer in the 'flow', but are absolutely positioned
which I now think looks better.  The text details div also now has a
white semitransparent background so that the graph can slide under it
and not look weird or mess up the text.  Better use of screen
real-estate, imho.

Fixes #30.